Move cert manager tests to correct file
